+++ packages/codefixes/README.md
@@ -1,3 +1,3 @@
-# @rehearsal/transforms
+# @rehearsal/codefixes
 
 Rehearsal Dependency Transforms Collection based on [TSC Diagnostic](https://github.com/microsoft/TypeScript/blob/main/src/compiler/diagnosticMessages.json)
