+++ packages/codefixes/src/2322/fixtures/2322.ts.output
