+++ packages/codefixes/src/2345/fixtures/2345.ts.output
