+++ packages/codefixes/src/2571/fixtures/2571.ts.output
