+++ packages/codefixes/src/7006/fixtures/7006.ts.input
