+++ packages/plugins/src/plugins/discover-types.plugin.ts
@@ -1,6 +1,6 @@
 import { Plugin, type PluginParams, type PluginResult } from '@rehearsal/service';
 import { findNodeAtPosition } from '@rehearsal/utils';
-import { codefixes, FixTransform } from '@rehearsal/transforms';
+import { codefixes, FixTransform } from '@rehearsal/codefixes';
 
 import { getFilesData } from '../data';
